+++ R/interactiveTable.R
@@ -232,7 +232,7 @@ print.interactiveTable <- function(x, useViewer, ...) {
     htmlPage <- paste("<html>",
       "<head>",
       "<meta http-equiv=\"Content-type\" content=\"text/html; charset=UTF-8\">",
-      "<script src=\"https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js\"></script>",
+      "<script src=\"https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js\"></script>",
       "</head>",
       "<body>",
       "<div style=\"margin: 0 auto; display: table; margin-top: 1em;\">",