+++ NEWS.md
@@ -7,6 +7,7 @@ Changes for 2.4.0
   the entire input `data.frame` with rounded elements as requested.
 * The `txtMergeLines` now returns an `knitr::asis_output` if html mode is activated. Fixed also
   multiline strings conversion into latex/html format.
+* Added vignette for the text formatting functions.
 
 Changes for 2.3.0
 -----------------